+++ ChangeLog.md
@@ -1,6 +1,7 @@
 ## change log ##
 
 * 2016-03-06: 0.3.1:
+  - add basic test for "self-hosted" request
   - add tests for requested formats
   
 * 2016-03-03: 0.3.1: